server release refactor
