+++ src/components/tables/CartProductsTable.vue
@@ -54,7 +54,6 @@
               size="14px"
               color="negative"
               dense
-              unelevated
               round
             ></q-btn>
           </q-td>
